server/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-redis==5.4.0
psycopg[binary]==3.2.10
redis==5.0.8
PyNaCl==1.5.0
cryptography==43.0.3
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
psutil==5.9.6
requests==2.31.0
channels==4.3.1
channels-redis==4.3.0
PyJWT==2.10.1
Pillow==11.3.0
